fix: switching between parcels and estate details (#549)
Closes #410 

This PR combines the `EstateDetailPage` and `ParcelDetailPage` under a parent `AssetDetailPage`. This removes some duplicated logic, and most importantly it prevents the `AssetPreviewHeader` (fka AssetDetailPage) to remount when switching between parcel and estate detail pages, preserving the zoom and avoiding the flickering.

I also took this chance to split the EstateEditPage from the EstateDetailPage, since those are very different components with their own logic and purpose.