fix: remove "Waiting for confirmations" to prevent flickering (#562)
Fixes #561 (sort of)

Since we don't store the number of confirmations in the app state, we always have to fetch them to know if a transaction's status is `Accepted` or `N confirmations`. 

I removed the `Waiting for confirmations...` placeholder that shows while we fetch the confirmations. This causes the status to just "appear" when we know the amount of confirmations, but I think it's better than showing the placeholder for a fraction of a second.

Any better ideas?