feat: defer update of the owner, update owners of selection, remove sold parcels from actual purchase (#710)
This PR does a bunch of things:

+ defers the update of the owner of the selected parcel so it can be rendered as selected right away

+ refresh the owners of the selected parcels every 10 seconds (configurable)

+ removes sold parcels from the actual purchase (they are deducted from the total of parcels and the total price)

As you can see in this screen recording, the user can select parcels right away, and some of them change to "sold" status after selecting them. They are not accounted for the total price and total LAND amount. They still show up on the list to let the user know that those parcels have been sold, and the user can click on them to remove them from the list (either by clicking the X on the corner of the chip, or by clicking on the tile on the map). Once they are removed from the list, the user can't add them again by clicking on them.

![689b](https://user-images.githubusercontent.com/2781777/49035871-d2452200-f194-11e8-83c5-2119b4b4632f.gif)

I also added small fixes for some issues:

+ Updated the colors of the SVGs in the sign in page

+ Fixed the total price shown in the confirmation modal

+ Updated the copy for "parcels available" as requested by ari.m

Closes #689